适配 PaddlePaddle 3.x 并新增各章虚拟测试
兼容性修复（13 文件，26 处替换）：
- `.numpy()[0]` → `.item()`：0-d 张量在 Paddle 2.5+ 起索引会抛 IndexError
- `preds==labels` / `paddle.equal(preds, labels)` → 比较前 cast labels 到
  preds.dtype：Paddle 3.x 移除了 float32↔int64 / int32↔int64 的隐式类型提升

新增 tests/ 目录：每章一个虚拟测试程序，共 8 章 48 用例，全部通过：
- chap1: 张量操作 + 自定义 Op 链 (add/multiply/exponential) 正反向
- chap2: 闭式最小二乘还原 w/b、L2 ridge、多项式特征
- chap3: model_SR 前向+反向、MultiCE 数值验证、iris 一步 SGD
- chap4: MLP loss 下降、RunnerV3 在 iris 完整 3 epoch 训练
- chap5: 自定义 Conv2D/Pool2D、Paddle_LeNet 一步训练、ResBlock、ResNet18 反向
- chap6: 自定义 SRN 训练、SimpleRNN/双向 LSTM、Embedding+LSTM 序列分类
- chap7: SGD/Adam/AdamW 对比、BN/LN/Dropout、L2Decay、StepDecay
- chap8: 缩放点积注意力 + 因果掩码、MHA、Embedding+MHA 分类训练

新增 .gitignore：忽略 .venv*, __pycache__, *.pdparams 等

测试环境：Python 3.12.10 + PaddlePaddle 3.3.1

+++ chap3线性模型/nndl/metric.py
@@ -17,4 +17,4 @@ def accuracy(preds, labels):
     else:
         #多分类时，使用'paddle.argmax'计算最大元素索引作为类别
         preds = paddle.argmax(preds,axis=1, dtype='int32')
-    return paddle.mean(paddle.cast(paddle.equal(preds, labels),dtype='float32'))
+    return paddle.mean(paddle.cast(paddle.equal(preds, paddle.cast(labels, dtype=preds.dtype)),dtype='float32'))
